# revise multi reductions per block in layer norm backward