--- a/benchmark/layer_norm_backward.cpp
+++ b/benchmark/layer_norm_backward.cpp
@@ -347,6 +347,25 @@ NVFUSER_BENCHMARK_RUN(NvFuserScheduler_LayerNorm_BWD_non64_fp32)
     ->Unit(benchmark::kMicrosecond)
     ->UseManualTime();
 
+// Multi-reductions per block
+NVFUSER_BENCHMARK_DEFINE(
+    NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16,
+    setupLayerNorm_BWD,
+    NvFuserScheduler_LayerNorm_BWD,
+    DataType::Half);
+NVFUSER_BENCHMARK_RUN(NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16)
+    ->Args({8 * 1024, 128})
+    ->Args({8 * 1024, 256})
+    ->Args({8 * 1024, 384})
+    ->Args({8 * 1024, 512})
+    ->Args({8 * 1024, 640})
+    ->Args({8 * 1024, 768})
+    ->Args({8 * 1024, 896})
+    ->Args({8 * 1024, 1023})
+    ->Args({8 * 1024, 1024})
+    ->Unit(benchmark::kMicrosecond)
+    ->UseManualTime();
+
 //------------------------------------------------------------------------------
 
 BENCHMARK(Baseline_LayerNorm_BWD_fp32)