Simple change to prioritize multi inner reductions per block. Performance increased due to the use of larger block sizes. e.g. previously there are only 16 threads per CTA for case with hidden size of 128. With revision, it changed to 512 and the performance increased from 43 GB/s to 131 GB/s.
Old performance on A100.
```
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                                                                                      Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/128/manual_time         135 us          230 us          102 bytes_per_second=43.5671G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/1/16)/grid(1/108/1)/64]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/256/manual_time        74.6 us          164 us          188 bytes_per_second=157.534G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: pad to warp / persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/2/32)/grid(1/108/1)/256]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/384/manual_time        87.7 us          179 us          160 bytes_per_second=200.868G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/2/48)/grid(1/108/1)/384]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/512/manual_time        64.3 us          153 us          218 bytes_per_second=365.208G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: pad to warp / persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/3/64)/grid(1/108/1)/768]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/640/manual_time        82.1 us          173 us          171 bytes_per_second=357.241G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/3/80)/grid(1/108/1)/960]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/768/manual_time        60.5 us          148 us          231 bytes_per_second=581.942G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: pad to warp / persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/4/96)/grid(1/108/1)/1536]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/896/manual_time        82.0 us          173 us          170 bytes_per_second=500.587G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/4/112)/grid(1/108/1)/1792]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/1023/manual_time       88.6 us          176 us          160 bytes_per_second=528.781G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: pad to warp / persistent batch - 4 / /Launch_Parameters[block(1/2/256)/grid(1/216/1)/2048]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/1024/manual_time       69.0 us          158 us          203 bytes_per_second=680.145G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: pad to warp / persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/4/128)/grid(1/108/1)/2048]
```

New performance on A100
```
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                                                                                      Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/128/manual_time        44.8 us          130 us          311 bytes_per_second=131.472G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/32/16)/grid(1/108/1)/2048]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/256/manual_time        44.4 us          130 us          315 bytes_per_second=264.653G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: pad to warp / persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/16/32)/grid(1/108/1)/2048]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/384/manual_time        46.4 us          134 us          301 bytes_per_second=379.735G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/10/48)/grid(1/108/1)/1920]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/512/manual_time        48.3 us          133 us          289 bytes_per_second=486.21G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: pad to warp / persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/8/64)/grid(1/108/1)/2048]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/640/manual_time        61.7 us          150 us          227 bytes_per_second=475.029G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/6/80)/grid(1/108/1)/1920]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/768/manual_time        57.5 us          145 us          244 bytes_per_second=612.45G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: pad to warp / persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/5/96)/grid(1/108/1)/1920]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/896/manual_time        82.5 us          173 us          171 bytes_per_second=497.797G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/4/112)/grid(1/108/1)/1792]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/1023/manual_time       88.3 us          176 us          158 bytes_per_second=531.031G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: pad to warp / persistent batch - 4 / /Launch_Parameters[block(1/2/256)/grid(1/216/1)/2048]
NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16___GRAPH/NvFuserScheduler_LayerNorm_BWD_MultiReductionsPerBlock_fp16/8192/1024/manual_time       69.0 us          160 us          203 bytes_per_second=679.989G/s Red On Fastest Dim // Persistent Kernel // Project Persistent Buffers //  // Iteration Domain: multiple reductions per block / split grid dimension /  // Inner Reduction Domain: pad to warp / persistent batch - 1 / vectorize / factor 8/Launch_Parameters[block(1/4/128)/grid(1/108/1)/2048]
```
![image](https://github.com/NVIDIA/Fuser/assets/116412316/57cde660-8a7f-48e1-8d39-5c2d5df72340)